requirements.txt
# ======================================================
# Judo Academy - Dependências Python
# ======================================================

# Framework Web
Flask==3.0.0
Werkzeug==3.0.1

# Autenticação e Sessões
Flask-Login==0.6.3

# Banco de Dados MySQL
mysql-connector-python==8.2.0

# Manipulação de Datas
python-dateutil==2.8.2

# Sync currículo Zempo (fetch + parse HTML)
requests==2.31.0
beautifulsoup4==4.12.2

# Leitura de PDF (importar calendário)
pypdf==4.0.1
pdfplumber==0.11.4

# Geração de Excel (XLSX)
openpyxl==3.1.2

# ======================================================
# Bibliotecas Padrão (não precisam ser instaladas)
# ======================================================
# - datetime
# - os
# - base64
# - unicodedata
# - json
# - hashlib

# ======================================================
# Instalação:
# ======================================================
# pip install -r requirements.txt
#
# Ou instalar individualmente:
# pip install Flask Flask-Login mysql-connector-python python-dateutil
